+++ HISTORY.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- Sets the default protocol to docker.
+  [href]
+
 # 0.0.0 (2018-08-24)
 
 - Initial Release.